Support gamepads and configurable joystick buttons (#1)
Previously, any joystick button would fire weapons. This adds configurable button indices for fire and self-destruct.

Added SDL game controller support. This is helpful for the included abstractions. D-pad to move, A button or R1 to fire, Y button to self-destruct. D-pad and A also work for the menu, and Start to pause/unpause.